observatorio-delito-app/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
pydantic==2.9.2
python-dotenv==1.0.1
pandas==2.2.2
openpyxl==3.1.5
geoalchemy2==0.15.2